% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{$,Nonpareil.Curve-method}
\alias{$,Nonpareil.Curve-method}
\title{Get attribute.}
\usage{
\S4method{$}{Nonpareil.Curve}(x, name)
}
\arguments{
\item{x}{\code{Nonpareil.Curve} object.}

\item{name}{Attribute.}
}
\description{
Get attribute.
}
