% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\docType{methods}
\name{$,Nonpareil.Set-method}
\alias{$,Nonpareil.Set-method}
\title{Get attribute.}
\usage{
\S4method{$}{Nonpareil.Set}(x, name)
}
\arguments{
\item{x}{\code{Nonpareil.Set} object.}

\item{name}{Attribute.}
}
\description{
Get attribute.
}
