% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\docType{methods}
\name{$<-,Nonpareil.Curve-method}
\alias{$<-,Nonpareil.Curve-method}
\title{Set attribute.}
\usage{
\S4method{$}{Nonpareil.Curve}(x, name) <- value
}
\arguments{
\item{x}{\code{Nonpareil.Curve} object.}

\item{name}{Attribute.}

\item{value}{New value.}
}
\description{
Set attribute.
}
