% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{$<-,Nonpareil.Set-method}
\alias{$<-,Nonpareil.Set-method}
\title{Set attribute.}
\usage{
\S4method{$}{Nonpareil.Set}(x, name) <- value
}
\arguments{
\item{x}{\code{Nonpareil.Set} object.}

\item{name}{Attribute.}

\item{value}{New value.}
}
\description{
Set attribute.
}
