\name{Nonpareil-package}
\alias{Nonpareil-package}
\alias{Nonpareil}
\docType{package}
\title{Ancilliary R Code for Nonpareil}
\description{A collection of functions to plot, process and analyze
files produced by Nonpareil (http://enve-omics.ce.gatech.edu/nonpareil).}
\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}

\keyword{ package }



