\name{Nonpareil.add.curve}
\alias{Nonpareil.add.curve}
\title{Nonpareil add curve}
\description{Adds a `Nonpareil.Curve` to a `Nonpareil.Set`}
\usage{Nonpareil.add.curve(nps, np)}
\arguments{
  \item{nps}{`Nonpareil.Set` object}
  \item{np}{`Nonpareil.Curve` object}
}



\author{Luis M. Rodriguez-R [aut, cre]}





