\name{Nonpareil.antif}
\alias{Nonpareil.antif}
\title{Nonpareil antif}
\description{Complement function of `Nonpareil.f`}
\usage{Nonpareil.antif(y, a, b)}
\arguments{
  \item{y}{Values of abundance-weighted average coverage}
  \item{a}{Parameter alpha of the gamma CDF.}
  \item{b}{Parameter beta of the gamma CDF.}
}

\value{Estimated sequencing effort.}

\author{Luis M. Rodriguez-R [aut, cre]}





