\name{Nonpareil.col}
\alias{Nonpareil.col}
\title{Nonpareil col}
\description{Returns the color of the curve}
\usage{Nonpareil.col(x, alpha = 1)}
\arguments{
  \item{x}{`Nonpareil.Curve` or `Nonpareil.Set` object}
  \item{alpha}{Alpha level of the color from 0 to 1}
}



\author{Luis M. Rodriguez-R [aut, cre]}





