\name{Nonpareil.coverage_factor}
\alias{Nonpareil.coverage_factor}
\title{Nonpareil coverage factor}
\description{Factor to transform redundancy into coverage (internal function).}
\usage{Nonpareil.coverage_factor(x)}
\arguments{
  \item{x}{`Nonpareil.Curve` object}
}

\value{A numeric scalar.}

\author{Luis M. Rodriguez-R [aut, cre]}





