\name{Nonpareil.coverageFactor}
\alias{Nonpareil.coverageFactor}
\title{Nonpareil coverageFactor}
\description{Factor to transform redundancy into coverage (internal function).}
\usage{Nonpareil.coverageFactor(overlap)}
\arguments{
  \item{overlap}{Value of overlap (-L).}
}

\value{A numeric scalar.}

\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com>}





