\name{Nonpareil.f}
\alias{Nonpareil.f}
\title{Nonpareil f}
\description{Function of the projected model}
\usage{Nonpareil.f(x, a, b)}
\arguments{
  \item{x}{Values of sequencing effort (typically in bp)}
  \item{a}{Parameter alpha of the Gamma CDF}
  \item{b}{Parameter beta of the Gamma CDF}
}

\value{Predicted values of abundance-weighted average coverage.}

\author{Luis M. Rodriguez-R [aut, cre]}





