% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.fit_model}
\alias{Nonpareil.fit_model}
\title{Fit the sigmoidal model to the rarefied coverage.}
\usage{
Nonpareil.fit_model(np, weights.exp)
}
\arguments{
\item{np}{\code{Nonpareil.Curve} object.}

\item{weights.exp}{Numeric; see \code{Nonpareil.curve} for details.}
}
\description{
Fit the sigmoidal model to the rarefied coverage.
}
