% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.kappa_to_coverage}
\alias{Nonpareil.kappa_to_coverage}
\title{Transform values of kappa (redundancy) into coverage (internal function).}
\usage{
Nonpareil.kappa_to_coverage(x, kappa)
}
\arguments{
\item{x}{\code{Nonpareil.Curve} object.}

\item{kappa}{Values of kappa to use as basis for coverage estimates}
}
\value{
A numeric vector.
}
\description{
Transform values of kappa (redundancy) into coverage (internal function).
}
