\name{Nonpareil.legend}
\alias{Nonpareil.legend}
\title{Nonpareil legend}
\description{Generates a legend for Nonpareil plots, after calling Nonpareil.curve()
or Nonpareil.curve.batch().}
\usage{Nonpareil.legend(x = NULL, y = 0.3, ...)}
\arguments{
  \item{x}{X coordinate, or any character string accepted by legend (e.g.,
'bottomright').}
  \item{y}{Y coordinate.}
  \item{\dots}{Any other parameters supported by legend().}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com>}





