% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\docType{package}
\name{Nonpareil}
\alias{Nonpareil-package}
\alias{Nonpareil}
\title{Nonpareil: Metagenome Coverage Estimation and Projections for 'Nonpareil'.}
\description{
Plot, process, and analyze NPO files produced by 'Nonpareil'
\url{http://enve-omics.ce.gatech.edu/nonpareil/}.
}
\section{Citation}{

If you use Nonpareil, please cite:
Rodriguez-R et al. 2018. Nonpareil 3: Fast estimation of metagenomic coverage
and sequence diversity.
mSystems 3(3): e00039-18. DOI: 10.1128/mSystems.00039-18.

Rodriguez-R & Konstantinidis. 2014. Nonpareil: a redundancy-based approach to
assess the level of coverage in metagenomic datasets.
Bioinformatics 30 (5): 629-635. DOI: 10.1093/bioinformatics/btt584.

For an extended discussion on coverage in metagenomic data, see also:

Rodriguez-R & Konstantinidis. 2014. Estimating coverage in metagenomic data
sets and why it matters.
The ISME Journal 8: 2349–2351. DOI: 10.1038/ismej.2014.76.
}

\seealso{
Useful links:
\itemize{
  \item \url{http://enve-omics.ce.gatech.edu/nonpareil/}
}

}
\author{
\strong{Maintainer}: Luis M. Rodriguez-R \email{lmrodriguezr@gmail.com}

Authors:
\itemize{
  \item Luis M. Rodriguez-R \email{lmrodriguezr@gmail.com}
}

}
