\name{Nonpareil.read_data}
\alias{Nonpareil.read_data}
\title{Nonpareil read data}
\description{Read the data tables and extract direct estimates}
\usage{Nonpareil.read_data(x, correction.factor)}
\arguments{
  \item{x}{`Nonpareil.Curve` object}
  \item{correction.factor}{Logical; see `Nonpareil.curve` for details}
}



\author{Luis M. Rodriguez-R [aut, cre]}





