% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.read_metadata}
\alias{Nonpareil.read_metadata}
\title{Read the metadata headers.}
\usage{
Nonpareil.read_metadata(x)
}
\arguments{
\item{x}{\code{Nonpareil.Curve} object.}
}
\description{
Read the metadata headers.
}
