% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{+,Nonpareil.Set,ANY-method}
\alias{+,Nonpareil.Set,ANY-method}
\title{Alias of \code{Nonpareil.add.curve}.}
\usage{
\S4method{+}{Nonpareil.Set,ANY}(e1, e2)
}
\arguments{
\item{e1}{\code{Nonpareil.Set} object (\code{nps}).}

\item{e2}{\code{Nonpareil.Curve} object (\code{np}).}
}
\description{
Alias of \code{Nonpareil.add.curve}.
}
