\name{print.Nonpareil.Curve}
\alias{print.Nonpareil.Curve}
\title{print Nonpareil Curve}
\description{Prints and returns invisibly a summary of the `Nonpareil.Curve`
results}
\usage{\method{print}{Nonpareil.Curve}(x, ...)}
\arguments{
  \item{x}{`Nonpareil.Set` object}
  \item{\dots}{Additional parameters ignored}
}

\value{Returns the summary invisibly. See help for `summary.Nonpareil.Curve` for
additional information.}

\author{Luis M. Rodriguez-R [aut, cre]}





