% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{print.Nonpareil.Set}
\alias{print.Nonpareil.Set}
\title{Prints and returns invisibly a summary of the \code{Nonpareil.Set} results.}
\usage{
\method{print}{Nonpareil.Set}(x, ...)
}
\arguments{
\item{x}{\code{Nonpareil.Set} object.}

\item{...}{Additional parameters ignored.}
}
\value{
Returns the summary invisibly. See help for
\code{summary.Nonpareil.Curve} and \code{summary.Nonpareil.Set} for
additional information.
}
\description{
Prints and returns invisibly a summary of the \code{Nonpareil.Set} results.
}
