% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{summary.Nonpareil.Set}
\alias{summary.Nonpareil.Set}
\title{Returns a summary of the \code{Nonpareil.Set} results.}
\usage{
\method{summary}{Nonpareil.Set}(object, ...)
}
\arguments{
\item{object}{\code{Nonpareil.Set} object.}

\item{...}{Additional parameters ignored.}
}
\value{
Returns a matrix with different values for each dataset. For additional
details on the values returned, see help for
\code{summary.Nonpareil.Curve}.
}
\description{
Returns a summary of the \code{Nonpareil.Set} results.
}
