\name{Nonpareil-package}
\alias{Nonpareil-package}
\alias{Nonpareil}
\docType{package}
\title{Ancilliary R code for Nonpareil}
\description{A collection of functions to plot, process and analyze
files produced by Nonpareil (http://enve-omics.ce.gatech.edu/nonpareil).}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Nonpareil}
\packageIndices{Nonpareil}
}
\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com>}

Maintainer: \packageMaintainer{Nonpareil}

\keyword{ package }


