\name{Nonpareil-package}
\alias{Nonpareil-package}
\alias{Nonpareil}
\docType{package}
\title{Ancilliary R code for Nonpareil}
\description{A collection of functions to plot, process and analyze
files produced by Nonpareil (http://enve-omics.ce.gatech.edu/nonpareil).}
\details{
\tabular{ll}{Package: \tab Nonpareil\cr
Title: \tab Ancilliary R code for Nonpareil\cr
Type: \tab Package\cr
Version: \tab 3.2\cr
Author: \tab Luis M. Rodriguez-R <lmrodriguezr@gmail.com>\cr
Maintainer: \tab Luis M. Rodriguez-R <lmrodriguezr@gmail.com>\cr
Authors@R: \tab c(person("Luis M.","Rodriguez-R",role=c("aut","cre"),
email="lmrodriguezr@gmail.com"))\cr
URL: \tab http://enve-omics.ce.gatech.edu/nonpareil\cr
Depends: \tab R (>= 2.9)\cr
License: \tab Artistic-2\cr
LazyLoad: \tab yes\cr
Encoding: \tab UTF-8\cr}
}
\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com>}

\keyword{ package }


