\name{Nonpareil-package}
\alias{Nonpareil-package}
\alias{Nonpareil}
\docType{package}
\title{Metagenome Coverage Estimation and Projections for 'Nonpareil'}
\description{Plot, process, and analyze NPO files produced by
'Nonpareil' <http://enve-omics.ce.gatech.edu/nonpareil>.}
\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}

\keyword{ package }



