% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.add.curve}
\alias{Nonpareil.add.curve}
\title{Adds a \code{Nonpareil.Curve} to a \code{Nonpareil.Set}.}
\usage{
Nonpareil.add.curve(nps, np)
}
\arguments{
\item{nps}{\code{Nonpareil.Set} object.}

\item{np}{\code{Nonpareil.Curve} object.}
}
\value{
Returns the \code{Nonpareil.Set} including the newly added
\code{Nonpareil.Curve}.
}
\description{
Adds a \code{Nonpareil.Curve} to a \code{Nonpareil.Set}.
}
