% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.antif}
\alias{Nonpareil.antif}
\title{Complement function of \code{Nonpareil.f}.}
\usage{
Nonpareil.antif(y, a, b)
}
\arguments{
\item{y}{Values of abundance-weighted average coverage.}

\item{a}{Parameter alpha of the gamma CDF.}

\item{b}{Parameter beta of the gamma CDF.}
}
\value{
Estimated sequencing effort.
}
\description{
Complement function of \code{Nonpareil.f}.
}
