% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.col}
\alias{Nonpareil.col}
\title{Returns the color of the curve.}
\usage{
Nonpareil.col(x, alpha = 1)
}
\arguments{
\item{x}{\code{Nonpareil.Curve} or \code{Nonpareil.Set} object.}

\item{alpha}{Alpha level of the color from 0 to 1.}
}
\description{
Returns the color of the curve.
}
