% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.coverage_factor}
\alias{Nonpareil.coverage_factor}
\title{Factor to transform redundancy into coverage (internal function).}
\usage{
Nonpareil.coverage_factor(x)
}
\arguments{
\item{x}{\code{Nonpareil.Curve} object.}
}
\value{
A numeric scalar.
}
\description{
Factor to transform redundancy into coverage (internal function).
}
