\name{Nonpareil.curve.batch}
\alias{Nonpareil.curve.batch}
\title{Nonpareil curve batch}
\description{Generates a single plot (and a single table of estimates) with multiple Nonpareil
curves, from different .npo files.}
\usage{Nonpareil.curve.batch(files, overlap, r = NA, g = NA, b = NA, 
    libnames = NA, read.lengths = NA, col = NA, ...)}
\arguments{
  \item{files}{Vector of characters with the paths to the .npo files.}
  \item{overlap}{Value of the '-L' parameter (in Nonpareil, the default is 50). It can be a number (if all the curves
were generated with the same value) or a vector (in the same order of 'files'). See the 'overlap' value
of 'Nonpareil.curve()'. Use only with Nonpareil < v2.0.}
  \item{r}{Values of 'r' in 'Nonpareil.curve()' (red value).}
  \item{g}{Values of 'g' in 'Nonpareil.curve()' (green value).}
  \item{b}{Values of 'b' in 'Nonpareil.curve()' (blue value).}
  \item{libnames}{Vector of names of the libraries (corresponding to 'libname' in 'Nonpareil.curve()'). It must be characters,
not factors.}
  \item{read.lengths}{A vector of numbers indicating the length of the reads (corresponding to 'read.length' in 'Nonpareil.curve()').
Use only with Nonpareil < v2.0.}
  \item{col}{Values of 'col' in 'Nonpareil.curve()'.}
  \item{\dots}{Any other parameter accepted by 'Nonpareil.curve()' is supported.}
}

\value{A dataframe containing the values generated by 'Nonpareil.curve()'.}

\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com>}





