% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.curve.batch}
\alias{Nonpareil.curve.batch}
\title{Alias of \code{Nonpareil.set}.}
\usage{
Nonpareil.curve.batch(files, col = NA, labels = NA, plot = TRUE,
  plot.opts = list(), ...)
}
\arguments{
\item{files}{Vector with the paths to the .npo files.}

\item{col}{Color of the curves (vector). If not passed, values are randomly
assigned. Values are recycled.}

\item{labels}{Labels of the curves (vector). If not passed, values are determined by
the filename. Values are recycled.}

\item{plot}{If TRUE, it generates the Nonpareil curve plots.}

\item{plot.opts}{Any parameters accepted by \code{plot.Nonpareil.Set} as a list.}

\item{...}{Any additional parameters accepted by \code{Nonpareil.curve}.}
}
\description{
Alias of \code{Nonpareil.set}.
}
