% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.f}
\alias{Nonpareil.f}
\title{Function of the projected model.}
\usage{
Nonpareil.f(x, a, b)
}
\arguments{
\item{x}{Values of sequencing effort (in bp).}

\item{a}{Parameter alpha of the Gamma CDF.}

\item{b}{Parameter beta of the Gamma CDF.}
}
\value{
Predicted values of abundance-weighted average coverage.
}
\description{
Function of the projected model.
}
