\name{Nonpareil.fit_model}
\alias{Nonpareil.fit_model}
\title{Nonpareil fit model}
\description{Fit the sigmoidal model to the rarefied coverage}
\usage{Nonpareil.fit_model(np, weights.exp)}
\arguments{
  \item{np}{`Nonpareil.Curve` object}
  \item{weights.exp}{Numeric; See `Nonpareil.curve` for details}
}



\author{Luis M. Rodriguez-R [aut, cre]}





