\name{Nonpareil.legend}
\alias{Nonpareil.legend}
\title{Nonpareil legend}
\description{Generates a legend for Nonpareil plots}
\usage{Nonpareil.legend(np, x, y = 0.3, ...)}
\arguments{
  \item{np}{A `Nonpareil.Set` object or a list of `Nonpareil.Curve` objects}
  \item{x}{X coordinate, or any character string accepted by legend (e.g.,
'bottomright').}
  \item{y}{Y coordinate.}
  \item{\dots}{Any other parameters supported by legend().}
}

\value{Returns invisibly a list, same as `legend`}

\author{Luis M. Rodriguez-R [aut, cre]}





