% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.legend}
\alias{Nonpareil.legend}
\title{Generates a legend for Nonpareil plots.}
\usage{
Nonpareil.legend(np, x, y = 0.3, ...)
}
\arguments{
\item{np}{A \code{Nonpareil.Set} or \code{Nonpareil.Curve} object, or a list of
\code{Nonpareil.Curve} objects.}

\item{x}{X coordinate, or any character string accepted by legend (e.g.,
'bottomright').}

\item{y}{Y coordinate.}

\item{...}{Any other parameters supported by legend().}
}
\value{
Returns invisibly a list, same as \code{legend}.
}
\description{
Generates a legend for Nonpareil plots.
}
