% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{Nonpareil.read_data}
\alias{Nonpareil.read_data}
\title{Read the data tables and extract direct estimates.}
\usage{
Nonpareil.read_data(x, correction.factor)
}
\arguments{
\item{x}{\code{Nonpareil.Curve} object.}

\item{correction.factor}{Logical; see \code{Nonpareil.curve} for details.}
}
\description{
Read the data tables and extract direct estimates.
}
