% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\docType{class}
\name{Nonpareil.Set-class}
\alias{Nonpareil.Set-class}
\title{Collection of \code{Nonpareil.Curve} objects. This object can be produced by
\code{Nonpareil.curve.batch} and supports S4 methods \code{plot},
\code{summary}, and \code{print}.}
\description{
Collection of \code{Nonpareil.Curve} objects. This object can be produced by
\code{Nonpareil.curve.batch} and supports S4 methods \code{plot},
\code{summary}, and \code{print}.
}
\section{Slots}{

\describe{
\item{\code{np.curves}}{List of \code{Nonpareil.Curve} objects.}

\item{\code{call}}{Call producing this object.}
}}

