% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{plot.Nonpareil.Set}
\alias{plot.Nonpareil.Set}
\title{Plot a \code{Nonpareil.Set} object.}
\usage{
\method{plot}{Nonpareil.Set}(
  x,
  col = NA,
  labels = NA,
  main = "Nonpareil Curves",
  legend.opts = list(),
  ...
)
}
\arguments{
\item{x}{\code{Nonpareil.Set} object to plot.}

\item{col}{Color of the curves (vector).
If passed, it overrides the colors set in the \code{Nonpareil.Curve}
objects. Values are recycled.}

\item{labels}{Labels of the curves (vector). If passed, it overrides the labels set
in the \code{Nonpareil.Curve} objects. Values are recycled.}

\item{main}{Title of the plot.}

\item{legend.opts}{Any additional parameters passed to \code{Nonpareil.legend}.
If FALSE, the legend is not displayed.}

\item{...}{Any additional parameters passed to \code{plot.Nonpareil.Curve}.}
}
\value{
Returns invisibly a \code{Nonpareil.Set} object (same as \code{x} input).
}
\description{
Plot a \code{Nonpareil.Set} object.
}
