% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{predict.Nonpareil.Curve}
\alias{predict.Nonpareil.Curve}
\title{Predict the coverage for a given sequencing effort.}
\usage{
\method{predict}{Nonpareil.Curve}(object, lr = object$LR, ...)
}
\arguments{
\item{object}{\code{Nonpareil.Curve} object.}

\item{lr}{Sequencing effort for the prediction (in bp).}

\item{...}{Additional parameters ignored.}
}
\value{
Returns the expected coverage at the given sequencing effort.
}
\description{
Predict the coverage for a given sequencing effort.
}
