% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonpareil.R
\name{print.Nonpareil.Curve}
\alias{print.Nonpareil.Curve}
\title{Prints and returns invisibly a summary of the \code{Nonpareil.Curve} results.}
\usage{
\method{print}{Nonpareil.Curve}(x, ...)
}
\arguments{
\item{x}{\code{Nonpareil.Set} object.}

\item{...}{Additional parameters ignored.}
}
\value{
Returns the summary invisibly. See help for
\code{summary.Nonpareil.Curve} for additional information.
}
\description{
Prints and returns invisibly a summary of the \code{Nonpareil.Curve} results.
}
