\name{print.Nonpareil.Set}
\alias{print.Nonpareil.Set}
\title{print Nonpareil Set}
\description{Prints and returns invisibly a summary of the `Nonpareil.Set` results}
\usage{\method{print}{Nonpareil.Set}(x, ...)}
\arguments{
  \item{x}{`Nonpareil.Set` object}
  \item{\dots}{Additional parameters ignored}
}



\author{Luis M. Rodriguez-R [aut, cre]}





