\name{print.Nonpareil.Set}
\alias{print.Nonpareil.Set}
\title{print Nonpareil Set}
\description{Prints and returns invisibly a summary of the `Nonpareil.Set` results}
\usage{\method{print}{Nonpareil.Set}(x, ...)}
\arguments{
  \item{x}{`Nonpareil.Set` object}
  \item{\dots}{Additional parameters ignored}
}

\value{Returns the summary invisibly. See help for `summary.Nonpareil.Curve` and
`summary.Nonpareil.Set` for additional information}

\author{Luis M. Rodriguez-R [aut, cre]}





