\name{summary.Nonpareil.Curve}
\alias{summary.Nonpareil.Curve}
\title{summary Nonpareil Curve}
\description{Returns a summary of the Nonpareil.Curve results}
\usage{\method{summary}{Nonpareil.Curve}(object, ...)}
\arguments{
  \item{object}{`Nonpareil.Curve` object}
  \item{\dots}{Additional parameters ignored}
}



\author{Luis M. Rodriguez-R [aut, cre]}





