\name{summary.Nonpareil.Set}
\alias{summary.Nonpareil.Set}
\title{summary Nonpareil Set}
\description{Returns a summary of the Nonpareil.Set results}
\usage{\method{summary}{Nonpareil.Set}(object, ...)}
\arguments{
  \item{object}{`Nonpareil.Set` object}
  \item{\dots}{Additional parameters ignored}
}



\author{Luis M. Rodriguez-R [aut, cre]}





