\name{summary.Nonpareil.Set}
\alias{summary.Nonpareil.Set}
\title{summary Nonpareil Set}
\description{Returns a summary of the Nonpareil.Set results}
\usage{\method{summary}{Nonpareil.Set}(object, ...)}
\arguments{
  \item{object}{`Nonpareil.Set` object}
  \item{\dots}{Additional parameters ignored}
}

\value{Returns a matrix with different values for each dataset. For additional
details on the values returned, see help for `summary.Nonpareil.Curve`}

\author{Luis M. Rodriguez-R [aut, cre]}





